ingestion/requirements.txt
requests
python-dotenv
google-cloud-storage
google-cloud-firestore
google-auth